Move file_loading into io.nexus
