+++ src/scippneutron/io/nexus/_json_nexus.py
