+++ src/scippneutron/io/nexus/_nexus.py
