+++ src/scippneutron/io/nexus/load_nexus.py
